+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release --
 - ADDED: Additional Unit and Feature tests (#36)
 - FIXES: Badly formatted templates that were causing issues (#39)
 - FIXES: Entry fillable exposes `created_by_user_id` (#22)
+- FIXES: BotBlock Needs Session Storage of Field Name (#23)
 - CHANGED: Moved `app` logic to stand-alone composer package (#12)
 
 ### Version 0.0.1 (Alpha 1)